moved damage and hp calcs from ints to floats

+++ Assets/Scripts/Enemy Scripts/EnemyLife.cs
@@ -20,7 +20,7 @@ public class EnemyLife : CharacterStats
         
     }
 
-    public void Damage(int i, GameObject sender){
+    public void Damage(float i, GameObject sender){
         if (isDead)
             return;
         if (sender.layer == gameObject.layer)
