+++ Assets/Scripts/Manager Scripts/Stat.cs
@@ -6,23 +6,23 @@ using UnityEngine;
 public class Stat
 {
    
-   [SerializeField]private int baseValue;
+   [SerializeField]private float baseValue;
 
-   private List<int> modifiers = new List<int>();
+   private List<float> modifiers = new List<float>();
 
-   public int GetValue(){
-   int finalValue = baseValue;
+   public float GetValue(){
+   float finalValue = baseValue;
    modifiers.ForEach(x => finalValue += x);
     return finalValue;
    }
 
-   public void AddModifier(int modifier){
+   public void AddModifier(float modifier){
       if (modifier != 0){
          modifiers.Add(modifier);
       }
    }
 
-   public void RemoveModifier(int modifier){
+   public void RemoveModifier(float modifier){
       if (modifier != 0){
          modifiers.Remove(modifier);
       }
